Adds the possibility to customize CSS per view.
If the ?view= parameter is present in the URL, we will try to include
a stylesheet for view.css. This way each view can be customized independently
of the others. For example. buildiator/?view=big can have 4 columns while
buildiator/?view=small can have only 2.

Note that this will try linking the stylesheet even if it does not exist.